+++ ToothAndTail_MapTool/ToothAndTail_Client/CFox.cpp
@@ -77,6 +77,6 @@ void CFox::RegisterToRenderList(vector<CObj*>& _vecRenderList)
 void CFox::TakeDamage(float _fDamageAmount, CObj* _pAttacker)
 {
 	D3DXVECTOR3 vPos = GetXY();
-	GetGameWorld().GetListObjs().emplace_back(new CParticle_Blood(GetGameWorld(), vPos.x, vPos.y + 100.f, _fDamageAmount));
+	GetGameWorld().GetListObjs().emplace_back(new CParticle_Blood(GetGameWorld(), vPos.x, vPos.y + 100.f, static_cast<int>(_fDamageAmount)));
 	CComDepObj::TakeDamage(_fDamageAmount);
 }
